colocate href types in `href.ts`
