+++ packages/route-pattern/src/index.ts
@@ -3,13 +3,13 @@ export {
   type RoutePatternOptions,
   type RoutePatternMatch,
 } from './lib/route-pattern.ts'
-export type { Join, HrefArgs, Params } from './lib/types/index.ts'
+export type { Join, Params } from './lib/types/index.ts'
+export { ParseError } from './lib/errors.ts'
+
+// href
+export { type Args as HrefArgs, HrefError } from './lib/route-pattern/href.ts'
 
 // matchers
 export { type Matcher, type Match } from './lib/matcher.ts'
 export { ArrayMatcher } from './lib/array-matcher.ts'
 export { TrieMatcher } from './lib/trie-matcher.ts'
-
-// errors
-export { ParseError } from './lib/errors.ts'
-export { HrefError } from './lib/route-pattern/href.ts'
