--- packages/route-pattern/src/lib/types/href.ts
@@ -1,16 +1,0 @@
-import type { OptionalParams, RequiredParams } from './params'
-import type * as Href from '../route-pattern/href.ts'
-
-type ParamValue = string | number
-
-// prettier-ignore
-export type HrefArgs<source extends string> =
-  [RequiredParams<source>] extends [never] ?
-    [] | [null | undefined | Record<string, any>] | [null | undefined | Record<string, any>, Href.SearchParams] :
-    [HrefParams<source>, Href.SearchParams] | [HrefParams<source>]
-
-// prettier-ignore
-type HrefParams<source extends string> =
-  & Record<RequiredParams<source>, ParamValue>
-  & Partial<Record<OptionalParams<source>, ParamValue | null | undefined>>
-  & Record<string, unknown>
