+++ packages/route-pattern/src/lib/types/index.ts
@@ -1,3 +1,2 @@
-export type { HrefArgs } from './href.ts'
 export type { Join } from './join.ts'
 export type { Params } from './params.ts'
